GPGGA: Parse NMEA GPS GGA message
The GGA (global positioning system fix data) contains the following 17 fields:

01. Message ID, GGA protocol header
02. UTC Time, Time of fix. HHMMSS format
03. Latitude, Degrees + minutes
04. North/South hemisphere indicator
05. Longitude, Degrees + minutes
06. East/West indicator
07. Position Fix Status Indicator
08. Satellites Used, Range 0 to 12
09. HDOP, Horizontal Dilution of Precision
10. MSL Altitude
11. Units, Meters (fixed field)
12. Geoid Separation
13. Units, Meters (fixed field)
14. Age of Differential Corrections, Blank (Null) fields when DGPS is not used
15. Diff. Reference Station ID
16. CheckSum in Hex
17. <CR><LF>

The GGA message is represented by the NMEA::GPGGA class. This class stores the
    additional fields that are not present in the NMEA::NMEAData class:

    GPGGA:
    * time_t TimeStamp;
    * float Latitude;
    * float Longitude;
    * bool Status;
    * int SatiliteFixes;
    * float HDOP;
    * float MSL;
    * char uMSL;
    * float GeoidSeparation;
    * char uSep;
    * float DifferentialCorrectionAge;
    * float DifferentialStationID;

ASSUMPTIONS
===========

The GGA message does not pass any information about the date the data was
collected on. I am assuming the data was from the current day so it will use the
provided data from localtime() is correct.

Becuase of this change, I have added the overloaded function that only takes in
the time stamp for ParseTimeStamp and doesn't take a DateStamp.

ParseTimeStamp signatures:

* std::string ParseTimeStamp(std::string);
* std::string ParseTimeStamp(std::string, std::string);