Update message parser selection method
This is a terrible, terrible commit... Which also happens to make a major change
in the parsers architecture :D

Remove the case statement for selecting the parse method for processing the
actual message.

Replace it with a std::vector of pointers to MessageParsers. This array is
filled with structs which inherit from MessageParser.
e.g. struct GPRMCParser : MessageParsers

These are ordered in the vector using the NMEA_MESSAGE_TYPE enums. The correct
parser is then called like this:

Parsers[Result->Header->Type]->Parse(Result, Elements);

Parse will then assign the message struct to Result.